Split vector flexible function to vector alu.
